alienbook/requirements.txt
alembic==1.3.3
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
Pillow==7.2.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.1
WTForms==2.2.1
